clean trainer 1/n
