+++ src/pytorch_lightning/trainer/setup.py
@@ -1,0 +1,113 @@
+# Copyright The Lightning AI team.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""
+Houses the methods used to set up the Trainer.
+"""
+
+from typing import Optional, Union
+from pytorch_lightning.loggers.logger import DummyLogger
+from pytorch_lightning.utilities.exceptions import MisconfigurationException
+from pytorch_lightning.utilities.rank_zero import rank_zero_info
+
+
+def _init_debugging_flags(
+    trainer,
+    limit_train_batches: Optional[Union[int, float]],
+    limit_val_batches: Optional[Union[int, float]],
+    limit_test_batches: Optional[Union[int, float]],
+    limit_predict_batches: Optional[Union[int, float]],
+    fast_dev_run: Union[int, bool],
+    overfit_batches: Union[int, float],
+    val_check_interval: Optional[Union[int, float]],
+    num_sanity_val_steps: int,
+):
+    # init debugging flags
+    if isinstance(fast_dev_run, int) and (fast_dev_run < 0):
+        raise MisconfigurationException(
+            f"fast_dev_run={fast_dev_run!r} is not a valid configuration. It should be >= 0."
+        )
+    trainer.fast_dev_run = fast_dev_run
+
+    # set fast_dev_run=True when it is 1, used while logging
+    if fast_dev_run == 1:
+        trainer.fast_dev_run = True
+
+    trainer.overfit_batches = _determine_batch_limits(overfit_batches, "overfit_batches")
+    overfit_batches_enabled = overfit_batches > 0
+
+    if fast_dev_run:
+        num_batches = int(fast_dev_run)
+        if not overfit_batches_enabled:
+            trainer.limit_train_batches = num_batches
+            trainer.limit_val_batches = num_batches
+
+        trainer.limit_test_batches = num_batches
+        trainer.limit_predict_batches = num_batches
+        trainer.fit_loop.max_steps = num_batches
+        trainer.num_sanity_val_steps = 0
+        trainer.fit_loop.max_epochs = 1
+        trainer.val_check_interval = 1.0
+        trainer.check_val_every_n_epoch = 1
+        trainer.loggers = [DummyLogger()] if trainer.loggers else []
+        rank_zero_info(
+            f"Running in `fast_dev_run` mode: will run the requested loop using {num_batches} batch(es). "
+            "Logging and checkpointing is suppressed."
+        )
+    else:
+        if not overfit_batches_enabled:
+            trainer.limit_train_batches = _determine_batch_limits(limit_train_batches, "limit_train_batches")
+            trainer.limit_val_batches = _determine_batch_limits(limit_val_batches, "limit_val_batches")
+        trainer.limit_test_batches = _determine_batch_limits(limit_test_batches, "limit_test_batches")
+        trainer.limit_predict_batches = _determine_batch_limits(limit_predict_batches, "limit_predict_batches")
+        trainer.num_sanity_val_steps = float("inf") if num_sanity_val_steps == -1 else num_sanity_val_steps
+        trainer.val_check_interval = _determine_batch_limits(val_check_interval, "val_check_interval")
+
+    if overfit_batches_enabled:
+        trainer.limit_train_batches = overfit_batches
+        trainer.limit_val_batches = overfit_batches
+
+
+
+def _determine_batch_limits(batches: Optional[Union[int, float]], name: str) -> Union[int, float]:
+    if batches is None:
+        # batches is optional to know if the user passed a value so that we can show the above info messages only to the
+        # users that set a value explicitly
+        return 1.0
+
+    # differentiating based on the type can be error-prone for users. show a message describing the chosen behaviour
+    if isinstance(batches, int) and batches == 1:
+        if name == "limit_train_batches":
+            message = "1 batch per epoch will be used."
+        elif name == "val_check_interval":
+            message = "validation will run after every batch."
+        else:
+            message = "1 batch will be used."
+        rank_zero_info(f"`Trainer({name}=1)` was configured so {message}")
+    elif isinstance(batches, float) and batches == 1.0:
+        if name == "limit_train_batches":
+            message = "100% of the batches per epoch will be used."
+        elif name == "val_check_interval":
+            message = "validation will run at the end of the training epoch."
+        else:
+            message = "100% of the batches will be used."
+        rank_zero_info(f"`Trainer({name}=1.0)` was configured so {message}.")
+
+    if 0 <= batches <= 1:
+        return batches
+    if batches > 1 and batches % 1.0 == 0:
+        return int(batches)
+    raise MisconfigurationException(
+        f"You have passed invalid value {batches} for {name}, it has to be in [0.0, 1.0] or an int."
+    )
